Safari-611.1.10/LayoutTests/http/tests/xmlhttprequest/xhr-user-agent-expected.txt

Validates User-Agent header of XMLHttpRequest

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Synchronous XHR
PASS syncXHR.status is 200
PASS syncXHRUserAgent.includes('AppleWebKit/') is true

* Asynchronous XHR
PASS asyncXHR.status is 200
PASS asyncXHRUserAgent.includes('AppleWebKit/') is true

PASS syncXHRUserAgent === asyncXHRUserAgent is true

PASS successfullyParsed is true

TEST COMPLETE
